refactor: remove excessive types, and centralize modules
