+++ .eslintrc.json
@@ -107,7 +107,6 @@
       }
     ],
     "@typescript-eslint/no-redeclare": "error",
-    "@typescript-eslint/consistent-type-imports": "error",
     "@typescript-eslint/ban-types": "error",
     "@typescript-eslint/consistent-type-assertions": "error",
     "@typescript-eslint/consistent-type-definitions": "error",
@@ -118,7 +117,6 @@
     "@typescript-eslint/no-var-requires": "error",
     "@typescript-eslint/no-unnecessary-boolean-literal-compare": "error",
     "@typescript-eslint/no-unnecessary-condition": "error",
-    "@typescript-eslint/consistent-type-exports": "error",
     "@typescript-eslint/no-unsafe-assignment": "error",
     "@typescript-eslint/no-unsafe-call": "error",
     "@typescript-eslint/explicit-member-accessibility": "error",
