+++ .storybook/preview.ts
@@ -1,4 +1,4 @@
-import type { Preview } from '@storybook/react';
+import { Preview } from '@storybook/react';
 
 const preview: Preview = {
   parameters: {
