+++ src/components/base/select/Select.tsx
@@ -1,5 +1,4 @@
-import type { ReactNode } from 'react';
-import { useState } from 'react';
+import { ReactNode, useState } from 'react';
 
 interface ISelectProps {
   children: ReactNode;
