+++ src/components/layouts/DailyHighlightCarousel/data/randomPhrase.ts
