+++ src/components/layouts/Notify/mappers/index.ts
@@ -1,5 +1,4 @@
-import { NotifyFromApiMapper } from '@/mappers/notify/fromApi';
-import type { INotifyApi } from '@/mappers/notify/fromApi';
+import { INotifyApi, NotifyFromApiMapper } from '@/layouts/Notify/mappers/fromApi';
 
 export class NotifyMapper {
   public static fromApi(data: INotifyApi[]): NotifyFromApiMapper[] {
