+++ src/components/layouts/header/EnvironmentName.tsx
@@ -1,0 +1,9 @@
+import { envs } from '@/core/env';
+
+export const EnvironmentName = () => {
+  if (!envs.VITE_ENVIRONMENT || envs.VITE_ENVIRONMENT === 'production') {
+    return <div />;
+  }
+
+  return <div className="text-[2rem] font-[800] fixed top-1 left-4 z-[190] text-white/75">{envs.VITE_ENVIRONMENT}</div>;
+};
