+++ src/components/widgets/cards/card.tsx
@@ -3,7 +3,7 @@ import { AiOutlineEye } from 'react-icons/ai';
 import { FaDivide, FaPercent } from 'react-icons/fa';
 import { BiCommentDetail, BiLike } from 'react-icons/bi';
 import { ReactionEnum } from '@/services/ReactionsService';
-import type { IMusicWithTransformation } from '@/mappers/music/fromApi';
+import { IMusicWithTransformation } from '@/mappers/music/fromApi';
 
 interface ICardProps {
   playlistLocal: IMusicWithTransformation;
