+++ src/core/connections/features/currentScreen/selectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
-import type { RootState } from '@/connections/store';
-import type { ICurrentScreenFeature } from '@/connections/features/currentScreen/slices';
+import { RootState } from '@/connections/store';
+import { ICurrentScreenFeature } from '@/connections/features/currentScreen/slices';
 
 const currentScreenSelected = (state: RootState): ICurrentScreenFeature => state.features.currentScreenFeature;
