+++ src/core/connections/features/reactions/selectors.ts
@@ -1,6 +1,6 @@
 import { createSelector } from '@reduxjs/toolkit';
-import type { RootState } from '@/connections/store';
-import type { IReactionsFeature } from '@/connections/features/reactions/slices';
+import { RootState } from '@/connections/store';
+import { IReactionsFeature } from '@/connections/features/reactions/slices';
 
 const selectReactions = (state: RootState): IReactionsFeature => state.features.reactionsFeature;
