+++ src/core/mappers/music/index.ts
@@ -1,7 +1,5 @@
-import type { IMusicApi } from '@/mappers/music/fromApi';
-import { MusicFromApiMapper } from '@/mappers/music/fromApi';
-import type { IMusicToApiMapper } from '@/mappers/music/toApi';
-import { MusicToApiMapper } from '@/mappers/music/toApi';
+import { IMusicApi, MusicFromApiMapper } from '@/mappers/music/fromApi';
+import { IMusicToApiMapper, MusicToApiMapper } from '@/mappers/music/toApi';
 
 export class MusicMapper {
   public static fromApi(data: IMusicApi[]): MusicFromApiMapper[] {
