+++ src/core/services/api/Reload.ts
@@ -1,7 +1,7 @@
 import { dataMusic } from '@/data/data.reload';
-import { mockNotify } from '@/data/notify';
-import type { INotifyApi } from '@/mappers/notify/fromApi';
-import type { IMusicApi } from '@/core/mappers/music/fromApi';
+import { IMusicApi } from '@/core/mappers/music/fromApi';
+import { mockNotify } from '@/layouts/Notify/data/notify';
+import { INotifyApi } from '@/layouts/Notify/mappers/fromApi';
 
 export class ApiReload {
   // eslint-disable-next-line no-unused-vars, @typescript-eslint/no-unused-vars
